# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GradoEats Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, assigned, in_delivery, delivered, cancelled]
        - name: driverId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients with order stats
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
...
  /clients/{id}:
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        items:
          type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, assigned, in_delivery, delivered, cancelled]
        serviceType:
          type: string
          enum: [nourriture, taxi, confort, tabac, fleur]
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - customerName
        - customerPhone
        - deliveryAddress
        - items
        - totalAmount
        - status
        - serviceType
        - createdAt
        - updatedAt
    CreateOrderBody:
      type: object
      properties:
        orderNumber:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        items:
          type: string
        totalAmount:
          type: number
        serviceType:
          type: string
          enum: [nourriture, taxi, confort, tabac, fleur]
        notes:
          type: string
        sourceUrl:
          type: string
      required:
        - orderNumber
        - customerName
        - customerPhone
        - deliveryAddress
        - items
        - totalAmount
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, assigned, in_delivery, delivered, cancelled]
        serviceType:
          type: string
          enum: [nourriture, taxi, confort, tabac, fleur]
        driverId:
          type: integer
          nullable: true
        notes:
          type: string
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        isVip:
          type: boolean
        totalOrders:
          type: integer
        totalSpent:
          type: number
        lastOrderAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - isVip
        - totalOrders
        - totalSpent
        - createdAt
        - updatedAt
...
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
        isVip:
          type: boolean
      required:
        - name
        - phone
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
        isVip:
          type: boolean