# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradoEats Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyServiceType } from "./createOrderBodyServiceType";

export interface CreateOrderBody {
  orderNumber: string;
  customerName: string;
  customerPhone: string;
  deliveryAddress: string;
  items: string;
  totalAmount: number;
  serviceType?: CreateOrderBodyServiceType;
  notes?: string;
  sourceUrl?: string;
}