# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradoEats Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderServiceType } from "./orderServiceType";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  customerPhone: string;
  deliveryAddress: string;
  items: string;
  totalAmount: number;
  status: OrderStatus;
  serviceType: OrderServiceType;
  driverId?: number | null;
  driverName?: string | null;
  sourceUrl?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}