# lib/api-zod/src/generated/types/updateClientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradoEats Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateClientBody {
  name?: string;
  email?: string;
  address?: string;
  notes?: string;
  isVip?: boolean;
}