# lib/api-zod/src/generated/types/updateOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradoEats Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateOrderBodyServiceType } from "./updateOrderBodyServiceType";
import type { UpdateOrderBodyStatus } from "./updateOrderBodyStatus";

export interface UpdateOrderBody {
  status?: UpdateOrderBodyStatus;
  serviceType?: UpdateOrderBodyServiceType;
  driverId?: number | null;
  notes?: string;
}